frontend/requirements.txt
Flask>=2.3.2
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
gunicorn>=20.1.0
numpy>=1.24.0
